requirements_python37.txt
asgiref==3.6.0
Django
djangorestframework
gunicorn
#Django==3.2
#django-mssql-backend==2.8.1
#djangorestframework==3.13.1
#gunicorn==20.1.0
#mssql-django==1.2
pip-upgrade==0.0.6
pyodbc==4.0.39
python-decouple==3.5
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
Unipath==1.1
whitenoise==6.4.0
